chore(build): convert build scripts into Rslib plugins
